A graphical text scroller
This set of modules takes lines from a text file and scrolls them across the screen. It applies various effects to the text.

To keep it interesting, there is also a movin background of stars and a stream of stars coming from the centre of the screen.

This is meant as a learning project, while I get to terms with python.